refactor: extract replay test cli helpers
